src/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
confluent-kafka==2.7.0
distro==1.9.0
exceptiongroup==1.2.2
fastavro==1.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.0.2
openai==1.59.3
pandas==2.2.3
py4j==0.10.9.7
pydantic==2.10.4
pydantic_core==2.27.2
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
